# packages/contracts/openapi/agentic-delivery-api.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reason"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "AuditEventsResponse": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "AuditEventsResponse",
        "type": "object"
      },
...
      "Classification": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "enum": [
              "general",
              "integration",
              "security",
              "data"
            ],
            "title": "Category",
            "type": "string"
          },
          "complexity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Complexity",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "category",
          "complexity",
          "reason"
        ],
        "title": "Classification",
        "type": "object"
      },
      "CreateServiceRequest": {
        "additionalProperties": false,
        "properties": {
          "dataContext": {
            "$ref": "#/components/schemas/DataContextRequest"
          },
          "request": {
            "maxLength": 10000,
            "minLength": 10,
            "title": "Request",
            "type": "string"
          },
          "requestedLocale": {
            "default": "en",
            "maxLength": 35,
            "title": "Requestedlocale",
            "type": "string"
          }
        },
        "required": [
          "request"
        ],
        "title": "CreateServiceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "displayName",
          "roles"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ToolActionProposal": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "const": "create_delivery_ticket",
            "default": "create_delivery_ticket",
            "title": "Action",
            "type": "string"
          },
          "arguments": {
            "$ref": "#/components/schemas/TicketArguments"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "proposalDigest": {
            "title": "Proposaldigest",
            "type": "string"
          },
          "proposalId": {
            "title": "Proposalid",
            "type": "string"
          },
          "schemaHash": {
            "title": "Schemahash",
            "type": "string"
          }
        },
        "required": [
          "proposalId",
          "arguments",
          "schemaHash",
          "proposalDigest",
          "expiresAt"
        ],
        "title": "ToolActionProposal",
        "type": "object"
      },
...
      "WorkflowResponse": {
        "additionalProperties": false,
        "properties": {
          "actionProposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolActionProposal"
              },
              {
                "type": "null"
              }
            ]
          },
          "approvalDecision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "citations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Classification"
              },
              {
                "type": "null"
              }
            ]
          },
          "correlationId": {
            "title": "Correlationid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationGateResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "executionResult": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolExecutionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "governanceDecisions": {
            "items": {
              "$ref": "#/components/schemas/PolicyDecision"
            },
            "title": "Governancedecisions",
            "type": "array"
          },
          "guardrails": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GuardrailReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "modelPolicy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "processingContext": {
            "$ref": "#/components/schemas/ProcessingContext"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImplementationProposal"
              },
              {
                "type": "null"
              }
            ]
          },
          "request": {
            "title": "Request",
            "type": "string"
          },
          "requestedLocale": {
            "title": "Requestedlocale",
            "type": "string"
          },
          "retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "stateVersion": {
            "title": "Stateversion",
            "type": "integer"
          },
          "status": {
            "enum": [
              "received",
              "awaiting_approval",
              "approved",
              "rejected",
              "expired",
              "executing",
              "completed"
            ],
            "title": "Status",
            "type": "string"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/StateTransitionResponse"
            },
            "title": "Transitions",
            "type": "array"
          },
          "workflowId": {
            "title": "Workflowid",
            "type": "string"
          }
        },
        "required": [
          "workflowId",
          "status",
          "stateVersion",
          "createdAt",
          "correlationId",
          "request",
          "requestedLocale",
          "transitions",
          "processingContext",
          "governanceDecisions"
        ],
        "title": "WorkflowResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Agentic Delivery Reference API",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/local/tasks/process-next": {
      "post": {
        "operationId": "process_next_local_task_v1_local_tasks_process_next_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Process Next Local Task",
        "tags": [
          "local-worker"
        ]
      }
    },
    "/v1/operations/audit-events": {
      "get": {
        "operationId": "audit_events_v1_operations_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Events",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/v1/service-requests": {
      "post": {
        "operationId": "create_service_request_v1_service_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Service Request",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/session": {
      "get": {
        "operationId": "session_v1_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Session",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/workflows/{workflow_id}": {
      "get": {
        "operationId": "get_workflow_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/{workflow_id}/approval": {
      "post": {
        "operationId": "decide_workflow_approval_v1_workflows__workflow_id__approval_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Workflow Approval",
        "tags": [
          "approvals"
        ]
      }
    },
    "/v1/workflows/{workflow_id}/plan": {
      "post": {
        "operationId": "plan_workflow_v1_workflows__workflow_id__plan_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Plan Workflow",
        "tags": [
          "workflows"
        ]
      }
    }
  }
}